refactor: move common fixtures into conftest module
