+++ tests/conftest.py
@@ -1,5 +1,6 @@
 """common fixtures for use by all test classes"""
 
+import os.path
 import pytest
 import annif
 
@@ -14,3 +15,18 @@ def app():
 def app_with_initialize():
     app = annif.create_app(config_name='config.TestingInitializeConfig')
     return app
+
+
+@pytest.fixture(scope='module')
+def datadir(tmpdir_factory):
+    return tmpdir_factory.mktemp('data')
+
+
+@pytest.fixture(scope='module')
+def subject_corpus():
+    subjdir = os.path.join(
+        os.path.dirname(__file__),
+        'corpora',
+        'archaeology',
+        'subjects')
+    return annif.corpus.SubjectDirectory(subjdir)
